requirements.txt
asgiref==3.7.2
astroid==3.0.2
autopep8==2.0.4
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap5==2023.10
dill==0.3.7
django-crispy-forms==2.1
django-tinymce==3.6.1
idna==3.6
isort==5.13.2
mccabe==0.7.0
mysql-connector-python==8.2.0
mysqlclient==2.2.1
Pillow==10.1.0
platformdirs==4.1.0
protobuf==4.21.12
pycodestyle==2.11.1
pylint==3.0.3
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.3
urllib3==2.1.0
